# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.JsonErrorResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  common.JsonFailedValidationResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/common.ValidationError'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  common.JsonSuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  common.ValidationError:
    properties:
      condition:
        type: string
      error:
        type: string
      key:
        type: string
    type: object
  requests.BookRequest:
    properties:
      author:
        type: string
      description:
        type: string
      stock:
        type: integer
      title:
        type: string
    required:
    - author
    - stock
    - title
    type: object
  requests.BorrowRequest:
    properties:
      book_id:
        type: integer
      due_date:
        type: string
    required:
    - book_id
    - due_date
    type: object
  requests.GetDelBookRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  requests.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 4
        type: string
    required:
    - email
    - password
    type: object
  requests.RegisterRequest:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        minLength: 4
        type: string
      role:
        allOf:
        - $ref: '#/definitions/requests.Role'
        enum:
        - admin
        - member
    required:
    - email
    - firstname
    - lastname
    - password
    type: object
  requests.Role:
    enum:
    - admin
    - member
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleMember
  requests.UpdateBookRequest:
    properties:
      author:
        type: string
      description:
        type: string
      id:
        type: integer
      stock:
        type: integer
      title:
        type: string
    required:
    - id
    type: object
host: localhost:8080
info:
  contact:
    email: ryanwanjie1@gmail.com
    name: Wanjie-Ryan
    url: https://github.com/Wanjie-Ryan
  description: Library Management System API with JWT auth, Redis cache, and MySQL.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LMS API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user and returns access & refresh tokens
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: accessToken, refreshToken, user
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "400":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.JsonFailedValidationResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: Register payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "400":
          description: Invalid payload / user exists
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.JsonFailedValidationResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      summary: Register user
      tags:
      - Auth
  /books/create:
    post:
      consumes:
      - application/json
      description: Allows admin users to create a new book entry.
      parameters:
      - description: Book payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.BookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.JsonFailedValidationResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      security:
      - BearerAuth: []
      summary: Create book
      tags:
      - Books
  /books/delete:
    delete:
      consumes:
      - application/json
      description: Delete book by id in the payload
      parameters:
      - description: Delete Book payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.GetDelBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Book
      tags:
      - Books
  /books/getAll:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of books for members
      parameters:
      - default: "1"
        description: Page number
        in: query
        name: page
        type: string
      - default: "10"
        description: Limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      security:
      - BearerAuth: []
      summary: Get paginated books
      tags:
      - Borrow Books
  /books/getsingle:
    get:
      consumes:
      - application/json
      description: Get single book by id in the payload
      parameters:
      - description: Get Single Book payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.GetDelBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      security:
      - BearerAuth: []
      summary: Get single Books
      tags:
      - Books
  /books/update:
    patch:
      consumes:
      - application/json
      description: Updates Books
      parameters:
      - description: Update Book payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Books
      tags:
      - Books
  /borrow:
    post:
      consumes:
      - application/json
      description: Member is able to borrow books
      parameters:
      - description: Borrow Book payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.BorrowRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      security:
      - BearerAuth: []
      summary: Member can borrow books
      tags:
      - Borrow Books
  /filter:
    get:
      consumes:
      - application/json
      description: Filter book by date range
      parameters:
      - description: Start date in YYYY-MM-DD or YYYY-MM-DD HH:MM:SS
        in: query
        name: start_date
        required: true
        type: string
      - description: End date in YYYY-MM-DD or YYYY-MM-DD HH:MM:SS
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      summary: Filter Book
      tags:
      - Books
  /profile/lookup:
    get:
      description: Returns the current user's profile
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - Profile
  /search:
    get:
      consumes:
      - application/json
      description: Search book by author or title
      parameters:
      - description: Search query for title or author
        in: query
        name: search
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonSuccessResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/common.JsonErrorResponse'
      summary: Search Book
      tags:
      - Books
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"